refactor(#652): move test list pages to app router
