+++ frontend/src/app/(scoreboard)/tests/endgame/page.tsx
@@ -1,14 +1,14 @@
+import { ExamList } from '@/components/exams/ExamList';
+import { RookIcon } from '@/style/ChessIcons';
 import { ExamType } from '@jackstenglein/chess-dojo-common/src/database/exam';
 import { Container, Stack, Typography } from '@mui/material';
-import { RookIcon } from '../../style/ChessIcons';
-import { ExamList } from './ExamList';
 
 const COHORT_RANGES = ['0-1000', '1000-1500', '1500-2000', '2000+'];
 
 /**
  * Renders the Material > Tests > Endgame Tests page.
  */
-export const ListEndgameExamsPage = () => {
+export default function ListEndgameExamsPage() {
     return (
         <Container sx={{ py: 5 }}>
             <Stack spacing={4}>
@@ -23,4 +23,4 @@ export const ListEndgameExamsPage = () => {
             </Stack>
         </Container>
     );
-};
+}
