+++ frontend/src/app/(scoreboard)/tests/tactics/page.tsx
@@ -1,26 +1,26 @@
+import { ExamList } from '@/components/exams/ExamList';
+import { QueenIcon } from '@/style/ChessIcons';
 import { ExamType } from '@jackstenglein/chess-dojo-common/src/database/exam';
 import { Container, Stack, Typography } from '@mui/material';
-import { KingIcon } from '../../style/ChessIcons';
-import { ExamList } from './ExamList';
 
-const POLGAR_RANGES = ['0-500', '500-1000', '1000-1500', '1500+'];
+const TACTICS_RANGES = ['0-1000', '1000-1500', '1500-2000', '2000+'];
 
 /**
- * Renders the Material > Tests > Checkmate Tests page.
+ * Renders the Material > Tests > Tactics Tests page.
  */
-export const ListCheckmateExamsPage = () => {
+export default function ListTacticsExamsPage() {
     return (
         <Container sx={{ py: 5 }}>
             <Stack spacing={4}>
                 <Typography variant='h4'>
-                    <KingIcon
+                    <QueenIcon
                         fontSize='inherit'
                         sx={{ mr: 2, verticalAlign: 'center' }}
                     />
-                    Checkmate Tests
+                    Tactics Tests
                 </Typography>
-                <ExamList cohortRanges={POLGAR_RANGES} examType={ExamType.Polgar} />
+                <ExamList cohortRanges={TACTICS_RANGES} examType={ExamType.Tactics} />
             </Stack>
         </Container>
     );
-};
+}
